+++ lib/blocs/settings/speech/speech_settings_state.dart
@@ -20,6 +20,7 @@ class SpeechSettingsState with _$SpeechSettingsState {
   factory SpeechSettingsState({
     required Set<String> availableModels,
     required Set<String> downloadedModels,
+    required Map<String, double> downloadProgress,
     String? selectedModel,
   }) = _SpeechSettingsState;
 }